+++ fabric/src/main/java/com/lx862/jcm/mod/JCMClient.java
@@ -1,6 +1,7 @@
 package com.lx862.jcm.mod;
 
-import com.lx862.jcm.mod.config.ClientConfig;
+import com.lx862.jcm.mod.config.JCMClientConfig;
+import com.lx862.jcm.mod.config.legacy.LegacyClientConfig;
 import com.lx862.jcm.mod.registry.JCMRegistryClient;
 import com.lx862.jcm.mod.render.gui.screen.ClientConfigScreen;
 import com.lx862.jcm.mod.resources.mcmeta.McMetaManager;
@@ -11,10 +12,10 @@ import org.mtr.mapping.holder.ScreenAbstractMapping;
 
 public class JCMClient {
     private static final McMetaManager mcMetaManager = new McMetaManager();
-    private static final ClientConfig config = new ClientConfig();
 
     public static void initializeClient() {
-        config.read();
+        new LegacyClientConfig().migrate();
+        JCMClientConfig.init();
         JCMRegistryClient.register();
         MTRScripting.register();
         JCMScripting.register();
@@ -22,10 +23,6 @@ public class JCMClient {
 
     public static ScreenAbstractMapping getClientConfigScreen(Screen previousScreen) {
         return new ClientConfigScreen().withPreviousScreen(previousScreen);
-    }
-
-    public static ClientConfig getConfig() {
-        return config;
     }
 
     public static McMetaManager getMcMetaManager() {