+++ fabric/src/main/java/com/lx862/jcm/mod/config/legacy/LegacyClientConfig.java
@@ -1,21 +1,22 @@
-package com.lx862.jcm.mod.config;
+package com.lx862.jcm.mod.config.legacy;
 
 import com.google.gson.JsonObject;
+import com.lx862.jcm.mod.config.JCMClientConfig;
 import com.lx862.jcm.mod.util.JCMLogger;
 import org.mtr.mapping.holder.MinecraftClient;
 
 import java.nio.file.Path;
 
-public class ClientConfig extends Config {
+@Deprecated
+public class LegacyClientConfig extends LegacyConfig {
 
     private static final Path CONFIG_PATH = MinecraftClient.getInstance().getRunDirectoryMapped().toPath().resolve("config").resolve("jsblock_client.json");
     public boolean disableRendering;
     public boolean debug;
-    public boolean showScriptLogSource;
     public boolean useNewTextRenderer;
     public boolean disableScriptingRestriction;
 
-    public ClientConfig() {
+    public LegacyClientConfig() {
         super(CONFIG_PATH);
     }
 
@@ -26,7 +27,11 @@ public class ClientConfig extends Config {
         this.debug = jsonConfig.get("debug_mode").getAsBoolean();
         this.useNewTextRenderer = jsonConfig.get("new_text_renderer").getAsBoolean();
         this.disableScriptingRestriction = jsonConfig.get("disable_scripting_restriction") != null && jsonConfig.get("disable_scripting_restriction").getAsBoolean();
-        this.showScriptLogSource = jsonConfig.get("show_log_source").getAsBoolean();
+
+        JCMClientConfig.INSTANCE.disableRendering.setValue(this.disableRendering);
+        JCMClientConfig.INSTANCE.debugMode.setValue(this.debug);
+        JCMClientConfig.INSTANCE.scripting.scriptDebugMode.setValue(this.debug);
+        JCMClientConfig.INSTANCE.scripting.disableScriptRestrictions.setValue(this.disableScriptingRestriction);
     }
 
     @Override
@@ -37,11 +42,6 @@ public class ClientConfig extends Config {
         jsonConfig.addProperty("debug_mode", debug);
         jsonConfig.addProperty("new_text_renderer", useNewTextRenderer);
         jsonConfig.addProperty("disable_scripting_restriction", disableScriptingRestriction);
-        jsonConfig.addProperty("show_log_source", showScriptLogSource);
         return jsonConfig;
-    }
-
-    public boolean newTextRenderer() {
-        return this.useNewTextRenderer;
     }
 }